Simplify SSDT-CPU.aml to rely on sole HWP
As noted in https://www.insanelymac.com/forum/topic/321021-guide-hwpintel-speed-shift-enable-with-full-power-management/?do=findComment&comment=2404269
the injection of 0x01 for plugin-type is enough to make the X86*.kext's
to load. Geekbench performance and power consumption seem unaffected by
this simplification.

I experimented with EPP changes and chose to stick to default value 0x80
previously used, but one can change that (cf explanations in .dsl file).