Add shell option for accessing hidden settings
The OpenCore documentation recommends that the setup allows macOS to
unlock the 0xE2 registry. My configuration requires this to boot (I
haven't tested the `AppleCpuPmCfgLock` and `AppleXcpmCfgLock` quirks).

As a `CFG Lock` option is available in the firmware, but is hidden from
the BIOS GUI, I included a GRUB shell modified by 'datasone' to access
this option.

Therefore, for this motherboard and BIOS F20 or F22f, you will have to
access this GRUB shell through the initial menu, type the command
`setup_var 0x4FE 0x00`, and reboot afterwards.
For other motherboards or other version of BIOSes, the OpenCore
documentation explains how to find the correct value.