Remove USB SSDT & replace w/ properties in kext
- For port discovery, corpnewt's USBMap (with XhciPortLimit quirk
temporarily enabled) does a better job than hackintool + USBInjectAll,
but hackintool helps for adding a few useful properties in the kext.
- I verified in System Report that iPhone connected through USB reveals
USB power properties
- As a convenience, I also modified my choice of port (it's easier for
me to remember that one of the front panel port is USB3 only, than
trying to remember which port at the back panel does not work).
- Also fixed HS09 as being with an USB2 connector (type 0).

https://github.com/headkaze/Hackintool
https://github.com/corpnewt/USBMap
https://www.tonymacx86.com/threads/the-new-beginners-guide-to-usb-port-configuration.286553/